Only kill pid 1 if there are no other psql instances running

+++ images/conduit/postgres/shell-profile.sh
@@ -1,5 +1,7 @@
-
-
 psql "$(echo $DB_SECRET | jq -rc '"postgres://\(.username):\(.password)@\(.host):\(.port)/\(.dbname)"')"
 
-kill `ps -e -o pid,comm | grep 'tail' | awk '{print$1}'`
+if [[ `ps -e -o pid,comm | grep 'psql' | wc -l` == "0" ]]; then
+  kill `ps -e -o pid,comm | grep 'tail' | awk '{print$1}'`
+fi
+
+exit
